+++ common/src/main/java/net/irisshaders/iris/vertices/BlockSensitiveBufferBuilder.java
@@ -1,7 +1,7 @@
 package net.irisshaders.iris.vertices;
 
 public interface BlockSensitiveBufferBuilder {
-	void beginBlock(short block, short renderType, byte blockEmission, int localPosX, int localPosY, int localPosZ);
+	void beginBlock(int block, byte renderType, byte blockEmission, int localPosX, int localPosY, int localPosZ);
 
 	void endBlock();
 }
